samples: net: zperf: Convert to use usec directly
Currently the code uses hw cycles and tries to convert them
to usec. I noticed some failures with this, for example the test
duration was sometimes missed meaning that instead of testing 5
seconds, we bailed out after 2 sec etc. After the kernel k_timeout_t
changes, which added APIs to support usec accuracy, we can use usec
and ticks here. This simplifies the code a bit.

+++ samples/net/zperf/src/zperf_session.h
@@ -49,7 +49,7 @@ struct session {
 	uint32_t outorder;
 	uint32_t error;
 	uint64_t length;
-	uint32_t start_time;
+	int64_t start_time;
 	uint32_t last_time;
 	int32_t jitter;
 	int32_t last_transit_time;
